# api/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OZMap Test",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/regions/": {
            "get": {
                "summary": "Get all regions",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "operationId": "findAllRegions"
            },
            "post": {
                "summary": "Creates a region",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "location": {
                                        "type": "object",
                                        "properties": {
                                            "coordinates": {
                                                "type": "array",
                                                "items": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": {
                                                "type": "string",
                                                "enum": [
                                                    "Polygon"
                                                ]
                                            }
                                        }
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "user": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Resource created successfully",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "success": {
                                        "summary": "Exemplo de resposta de sucesso",
                                        "value": {
                                            "message": "Resource created successfully"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "error": {
                                        "value": {
                                            "message": "Resource not found"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "error": {
                                        "value": {
                                            "message": "Unprocessable Entity"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "error": {
                                        "value": {
                                            "message": "Internal Server error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "createRegion"
            }
        },
        "/api/v1/regions/{id}": {
            "get": {
                "summary": "Get region by ID",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "User ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "operationId": "getRegionById"
            },
            "patch": {
                "summary": "Updates a region",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "Region ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "location": {
                                        "type": "object",
                                        "properties": {
                                            "coordinates": {
                                                "type": "array",
                                                "items": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": {
                                                "type": "string",
                                                "enum": [
                                                    "Polygon"
                                                ]
                                            }
                                        }
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "user": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Resource created successfully",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "success": {
                                        "summary": "Exemplo de resposta de sucesso",
                                        "value": {
                                            "message": "Resource created successfully"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "error": {
                                        "value": {
                                            "message": "Resource not found"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "error": {
                                        "value": {
                                            "message": "Unprocessable Entity"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "error": {
                                        "value": {
                                            "message": "Internal Server error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "updateRegion"
            },
            "delete": {
                "summary": "Delete region by ID",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "description": "Region ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "operationId": "deleteRegionById"
            }
        },
        "/api/v1/regions/point": {
            "post": {
                "summary": "List regions containing a specific point",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "lat": {
                                        "type": "number"
                                    },
                                    "lng": {
                                        "type": "number"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Resource created successfully",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "success": {
                                        "summary": "Exemplo de resposta de sucesso",
                                        "value": {
                                            "message": "Resource created successfully"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "error": {
                                        "value": {
                                            "message": "Resource not found"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "error": {
                                        "value": {
                                            "message": "Unprocessable Entity"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "error": {
                                        "value": {
                                            "message": "Internal Server error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "findRegionsContainingPoint"
            }
        },
...
    }
}